feat: implementation of the fast array (2nd part)
